refactor: reorganize training modules and improve input example handling
